refactor: update enum preference imports and deprecate old usage
