+++ generic-datastore/src/commonTest/kotlin/io/github/arthurkun/generic/datastore/preferences/core/AbstractDatastoreInstrumentedTest.kt
@@ -7,8 +7,8 @@ import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.core.stringSetPreferencesKey
 import io.github.arthurkun.generic.datastore.core.distinctFlow
 import io.github.arthurkun.generic.datastore.preferences.GenericPreferencesDatastore
-import io.github.arthurkun.generic.datastore.preferences.core.custom.enum
 import io.github.arthurkun.generic.datastore.preferences.core.customSet.enumSet
+import io.github.arthurkun.generic.datastore.preferences.enum
 import io.github.arthurkun.generic.datastore.preferences.toggle
 import io.github.arthurkun.generic.datastore.preferences.utils.map
 import io.github.arthurkun.generic.datastore.preferences.utils.mapIO
