+++ samples/preferenceComposeApp/src/commonMain/kotlin/io/github/arthurkun/generic/datastore/compose/app/domain/PreferenceStore.kt
@@ -2,8 +2,8 @@ package io.github.arthurkun.generic.datastore.compose.app.domain
 
 import io.github.arthurkun.generic.datastore.preferences.GenericPreferencesDatastore
 import io.github.arthurkun.generic.datastore.preferences.batch.BatchWriteScope
-import io.github.arthurkun.generic.datastore.preferences.core.custom.enum
 import io.github.arthurkun.generic.datastore.preferences.core.customSet.enumSet
+import io.github.arthurkun.generic.datastore.preferences.enum
 import io.github.arthurkun.generic.datastore.preferences.kserialized
 import io.github.arthurkun.generic.datastore.preferences.kserializedSet
 import io.github.arthurkun.generic.datastore.preferences.utils.mapIO
